Adding fileFormat to indicate which todo version was serialized
