+++ __tests__/__fixtures__/eslint-exact-matches.json
@@ -1,33 +1,6 @@
 {
   "results": [
     {
-      "filePath": "{{path}}/lint-app/.eslintrc.js",
-      "messages": [],
-      "errorCount": 0,
-      "warningCount": 0,
-      "fixableErrorCount": 0,
-      "fixableWarningCount": 0,
-      "usedDeprecatedRules": [{ "ruleId": "no-process-exit", "replacedBy": [] }]
-    },
-    {
-      "filePath": "{{path}}/lint-app/.template-lintrc.js",
-      "messages": [],
-      "errorCount": 0,
-      "warningCount": 0,
-      "fixableErrorCount": 0,
-      "fixableWarningCount": 0,
-      "usedDeprecatedRules": [{ "ruleId": "no-process-exit", "replacedBy": [] }]
-    },
-    {
-      "filePath": "{{path}}/lint-app/app/app.js",
-      "messages": [],
-      "errorCount": 0,
-      "warningCount": 0,
-      "fixableErrorCount": 0,
-      "fixableWarningCount": 0,
-      "usedDeprecatedRules": []
-    },
-    {
       "filePath": "{{path}}/lint-app/app/components/foo.js",
       "messages": [
         {
@@ -47,15 +20,6 @@
       "fixableErrorCount": 0,
       "fixableWarningCount": 0,
       "source": "import Component from \"@glimmer/component\";\n\nexport default class FooComponent extends Component {\n  constructor() {\n    super();\n\n    let foo = true;\n\n    console.log(\"bar\");\n  }\n}\n",
-      "usedDeprecatedRules": []
-    },
-    {
-      "filePath": "{{path}}/lint-app/app/router.js",
-      "messages": [],
-      "errorCount": 0,
-      "warningCount": 0,
-      "fixableErrorCount": 0,
-      "fixableWarningCount": 0,
       "usedDeprecatedRules": []
     },
     {
@@ -92,33 +56,6 @@
       "usedDeprecatedRules": []
     },
     {
-      "filePath": "{{path}}/lint-app/config/environment.js",
-      "messages": [],
-      "errorCount": 0,
-      "warningCount": 0,
-      "fixableErrorCount": 0,
-      "fixableWarningCount": 0,
-      "usedDeprecatedRules": [{ "ruleId": "no-process-exit", "replacedBy": [] }]
-    },
-    {
-      "filePath": "{{path}}/lint-app/config/targets.js",
-      "messages": [],
-      "errorCount": 0,
-      "warningCount": 0,
-      "fixableErrorCount": 0,
-      "fixableWarningCount": 0,
-      "usedDeprecatedRules": [{ "ruleId": "no-process-exit", "replacedBy": [] }]
-    },
-    {
-      "filePath": "{{path}}/lint-app/ember-cli-build.js",
-      "messages": [],
-      "errorCount": 0,
-      "warningCount": 0,
-      "fixableErrorCount": 0,
-      "fixableWarningCount": 0,
-      "usedDeprecatedRules": [{ "ruleId": "no-process-exit", "replacedBy": [] }]
-    },
-    {
       "filePath": "{{path}}/lint-app/rule-config.js",
       "messages": [
         {
@@ -138,42 +75,6 @@
       "fixableErrorCount": 0,
       "fixableWarningCount": 0,
       "source": "module.exports = {\n  'no-html-comments': { warn: 5, error: 10 },\n}\n",
-      "usedDeprecatedRules": []
-    },
-    {
-      "filePath": "{{path}}/lint-app/testem.js",
-      "messages": [],
-      "errorCount": 0,
-      "warningCount": 0,
-      "fixableErrorCount": 0,
-      "fixableWarningCount": 0,
-      "usedDeprecatedRules": [{ "ruleId": "no-process-exit", "replacedBy": [] }]
-    },
-    {
-      "filePath": "{{path}}/lint-app/tests/integration/components/bar-test.js",
-      "messages": [],
-      "errorCount": 0,
-      "warningCount": 0,
-      "fixableErrorCount": 0,
-      "fixableWarningCount": 0,
-      "usedDeprecatedRules": []
-    },
-    {
-      "filePath": "{{path}}/lint-app/tests/integration/components/foo-test.js",
-      "messages": [],
-      "errorCount": 0,
-      "warningCount": 0,
-      "fixableErrorCount": 0,
-      "fixableWarningCount": 0,
-      "usedDeprecatedRules": []
-    },
-    {
-      "filePath": "{{path}}/lint-app/tests/test-helper.js",
-      "messages": [],
-      "errorCount": 0,
-      "warningCount": 0,
-      "fixableErrorCount": 0,
-      "fixableWarningCount": 0,
       "usedDeprecatedRules": []
     }
   ]
