+++ __tests__/__fixtures__/eslint-no-fuzzy-source-prechange.json
@@ -6,7 +6,7 @@
         {
           "ruleId": "no-unused-vars",
           "severity": 2,
-          "message": "'bar1' is assigned a value but never used.",
+          "message": "'foo1' is assigned a value but never used.",
           "line": 7,
           "column": 9,
           "nodeType": "Identifier",
@@ -19,7 +19,7 @@
       "warningCount": 0,
       "fixableErrorCount": 0,
       "fixableWarningCount": 0,
-      "source": "import Component from \"@glimmer/component\";\n\nexport default class FooComponent extends Component {\n  constructor() {\n    super();\n\n    let bar1 = true;\n\n    console.log(\"bar\");\n  }\n}\n",
+      "source": "import Component from \"@glimmer/component\";\n\nexport default class FooComponent extends Component {\n  constructor() {\n    super();\n\n    let foo1 = true;\n\n    console.log(\"bar\");\n  }\n}\n",
       "usedDeprecatedRules": []
     },
     {
